Change name of mul language for terms in JS
Add getLanguageNameByCodeForTerms() to wikibase.getLanguageNameByCode,
similar to LanguageNameLookup::getNameForTerms() introduced in change
I7dbd94be13 (commit 4af1d0bfc3). Use it where appropriate (the use in
descriptionview is questionable, but doesn’t make a difference, because
the wikibase-description-edit-placeholder-not-applicable won’t use the
language name; I think it’s better to consistently use the “terms”
context here). Note that WikibaseLexeme also uses the “old” method, so
we don’t want it to return the “terms” language name all the time.

Reimplementing the mul logic in JS is a bit icky. An alternative would
be to have the ResourceLoader module (defined in ViewHooks) send the
language names from LanguageNameLookup to JS, but then we’d need to
adjust the logic to filter out language names from ULS, and somehow send
two language names for mul (one for “terms”, one for other contexts) in
the synthetic languageNames.json file. Reimplementing the logic in JS,
and sending the wikibase-language-name-for-terms-mul message as a
standard ResourceLoader message, seems like the lesser evil to me. (Add
some comments on both sides to point to the other implementation.)

In wikibase.WikibaseContentLanguages.js, make the constructor private
(as far as I can tell, nobody was calling it) and add a parameter. The
interface of any given WikibaseContentLanguages instance stays the same
(getAll(), getName(), getLanguageNameMap()), but the language name for
mul will depend on whether you got the instance from getTermLanguages()
or getMonolingualTextLanguages().

Bug: T345083
Change-Id: I03f0cfa9b04dc586556560fd2111c64d5930db8b