bridge: Upgrade Vue from v3.2.23 to v3.3.9
This follows the upgrade of the Vue version shipped with MediaWiki in
I0b3b6adf (commit ed6f7b2).
The goal is to have those two versions always exactly in sync.

Created with:
`npm install vue@3.3.9 --save-exact`

The inLanguagePlugin should never be called without a language code.
So if that ever happens, then there is a logic error somewhere in the
application, and directly throwing an Error is probably the most helpful
thing we can do to notice and fix it.

Bug: T352390
Change-Id: I5aaddaea969620b30af656fa20cf0fe1daac8c77