Use LanguageNameLookup in wbcontentlanguages API
Instead of using the underlying MediaWiki LanguageNameUtils, use our
custom LanguageNameLookup (via LanguageNameLookupFactory), and call
either getName() or getNameForTerms() based on the context parameter.

LanguageNameUtils returns empty string for unknown language names or
autonyms, and previously, we had some code to turn that into a nicer
null for API consumers, and to emit a warning if cldr was installed and
we still saw missing language names. But LanguageNameLookup instead
falls back to the language code in such cases. Comparing to the language
code, just to determine whether that fallback took place or not, feels
more brittle than comparing to the empty string; I think it’s better to
just return the string unmodified in that case. (If API users are
interested in whether MediaWiki core knows a certain language name or
not, they can use the languageinfo API instead; and as for cldr, I don’t
think it’s really our responsibility to integration-test that anyways.)
Removing the cldr integration also makes the factory method obsolete.

Bug: T345083
Change-Id: I25d1e9455741e30cb9c9c670fa9da016c728ac6f