Adjust box-sizing for headings on entity pages
By default, padding isn't included in the width, so setting the left
padding to 10px and the width to 100% means that these headings are
actually 100% + 10px wide.

This normally doesn't cause a problem because there is usually more
than 10px of extra space to the right of the heading. In the Timeless
skin, however, when the page width is below 850px, the padding on
#mw-content is 0.45em (= 6.84px), so the heading is too wide and
causes the page to scroll horizontally.

Changing the box-sizing to border-box causes the padding to be included
in the width, which seems like what was intended.

Bug: T335540
Bug: T330820
Change-Id: I52d0d123f84e967e1ead29aa4499b9faf9771b1f