Change anonymous edit warning based on tempuser configuration
For Mediawikis with support for Temporary Users, it is no longer the
case that IP addresses of anonymous editors will be shown. To support
this change, we need to fetch the TempUser configuration of the
repo to which wikidata-bridge is connecting and pass that config into
bridge to display an appropriate message.

Bug: T353807
Change-Id: I5ec257a87c2ee1871d9e4dd5ba10957a355bb4bd