Make addUsedEntities aware of statement grouping
addUsedEntities in SqlUsageTracker needs to be aware of statement
grouping so that it doesn't add a "Q1#C.P2" statement where a
"Q1#C" statement already exists, and likewise that it removes
"Q1#C.P2" statements when adding a "Q1#C" statement. This should
eliminate the rapid adding and removing of records in the usage
table that we observe in the bug ticket.

Bug: T255706
Change-Id: If47c4883cbc7077fa7c7bd168ae5d55b536dc4c0