REST: Add SiteLinkRetriever
This implements `SiteLinkRetriever::getSiteLink()` (singular) by getting
all sitelinks of the requested item and then finding the one for the
requested site in that list. Using `SiteLinkLookup::getLinks()` was
considered as a potentially more performant option, but it looks like it
might actually be slower in some cases considering that
`CachingSiteLinkLookup::getLinks()` bypasses the cache. This can be
optimized in the future if necessary.

Bug: T354332
Change-Id: I09bd3a7d7c575d00b4a019375b741b3659038de1