Fix secondary CI on GitHub by actually cloning MediaWiki
The new phpunit default configuration file is at MediaWiki's root repository
as of If2ffe052 (commit a69c57b). It is also, excluded from a export of
the mediawiki repository via the .gitattributes file.

For some unknown reason MediaWiki was not cloned but downloaded as an
archive file from GitHub, thus not including the phpunit config file,
thus failing to run the tests. By moving back to cloning the respective
branch of MediaWiki, the tests pass again.

Bug: T353957
Change-Id: Id9796f249705617d1ff310c6b50b8ec75a6d9662