Fix CI errors on REL1_39
This is a combination of several changes backported from the master
branch (the first three previously having been backported to REL1_40 as
change I34baed7374, commit 2ad7f59c40), which have to be merged together
to fix CI:

- TR: drop storybook (change Ib57537a6fd, bugs T345725, T345660)
- TR: fix dependencies (change I0b89c42261, bugs T345725, T345660)
- disable wdio tests (change I942b5dd688, bugs T293080, T256626)
- remove node16 hack (change I4e6462b7ac, bug T330391)
- TR: avoid unsafe regex (changes I81e7fdbb8d, I27b9df71b2)
- TR: skip test:distnodiff (change I2551801c80, bug T354881)

I then had to make further changes to the package{,-lock}.json files: I
removed all references to @wdio/sync and fibers, and somehow resolved
peer dependency conflicts in tainted references (which required
completely deleting and rebuilding the lockfile, hence the bump in the
lockfileVersion). So this commit is an absolute mess. If it makes CI
green, I’m happy.

Change-Id: Ic184f7f6a8170345a3b8e6a3d8a4b4a0421507cb