Only build result entries for used wbsearchentities results
Previously, SearchEntities would put together search match entries for
all the limit+continue search results (including looking up their page
ID), only to throw the first `continue` of them away again. Let’s do the
slicing first and only assemble the data for the results we care about.

We can test this by checking that the EntityArticleIdLookup service is
only called up to limit times. (My first idea was to have the
EntitySearchHelper return a mock TermSearchResult which would throw if
any of its methods were called, but that doesn’t work because
SearchEntities wraps the EntitySearchHelper in a ConceptUriSearchHelper,
which copies the TermSearchResult to a new object and in the process
calls all of its methods.)

Bug: T355053
Change-Id: Ibd8ab53ec3fac0fed49ad3ce290bb2b20e382576