Add limit to wbsearchentities continue parameter
Given how search with offset is implemented (just ask the backend for
continue+limit results and then slice off the first continue of them),
we want to put some limit on the continue parameter so we won’t ask for
arbitrarily many results from the search backend.

While we’re at it, also replace the deprecated self:: constants with
their IntegerDef:: versions in the definition of the limit parameter.

For the record, I’m not sure if the “soft limit” behavior was actually
intended – the comment about the “maximum continuation limit” to me kind
of sounds like it was meant to be a “hard limit”, but a check for that
wasn’t added. But 10000 results still seem to perform well enough, so
let’s just leave it as it is for the time being, IMHO.

Bug: T355251
Change-Id: Ib0ef82046456ef86c22e6e64df06e97284d86946