Fix bug in usage tracking for Parsoid parses
In the current implementation of Parsoid, it may be the case that
 `parser->resetOutput` (`Parsoid::Config::DataAccess`) gets called
without `parser->clearState` (`Parsoid::Parser`) being triggered.
Specifically, this is always the case for consecutive parses within
the context of the same PageReference.

`clearState` clears all the state and disposes of references to any
singletons as well as calling `resetOutput`, whereas `resetOutput`
only creates a new `ParserOutput` object.

In `Scribunto_LuaWikibaseEntityLibrary`, we were retaining a
reference to the `ParserOutput`, and in the case that `resetOutput`
is called outside of a `clearState`, the usage tracking was being
accumulated in a `ParserOutput` that the `Parser` no longer had a
reference to. This patch ensures that no stale reference is cached
in the `Scribunto_LuaWikibaseEntityLibrary`.

Because this behaviour only showed up with Parsoid, the bug
resulted in a difference in usage counts between Parsoid and
legacy parses, which we believe was causing occasional database
timeouts (see ticket).

Bug: T255706
Change-Id: I7410baa3e94fe04abaa2fd788d59c117f89f88e1