REST: Handle case sensitive statement IDs as HTTP redirect
Historically, Wikibase has allowed statement IDs starting with lowercase
Item or Property IDs. This resulted in mixed statement IDs on the same
Item, e.g. both q7348050$... and Q7348050$... As the statement ID is
used to identify it in REST API, but also in other Wikibase interfaces,
Wikibase REST API cannot stop handling those.

This change creates a 308 redirect response if there is a statement on
the requested Item or Property that matches the requested statement ID,
but with different case of character. The URL with the correct statement
ID will be provided as a value of the "Location:" header.

Bug: T354261
Change-Id: Ibcaaf325020f2971bc979b518f1d90944db6ffbc