Create temp users on redirect/merge
Make the EntityRedirectCreationInteractor create a temp user if needed
and return an appropriate DerivativeContext, just like
MediaWikiEditEntity; make ItemMergeInteractor thread these derivative
contexts through the successive edits (the temp user created by the
first edit should also be used by the other two edits); and have both
interactors return the temp user as well. The special pages and API
modules can then use the TempUserCreatedRedirect hook to generate the
right redirect URL and either redirect to it directly (special pages) or
include it in the response data (APIs).

In the special pages, also introduce a new “success” URL state that the
redirects can return to, with all the data that’s needed to display the
“success” message. (For Special:RedirectEntity, it’s just a boolean
flag; for Special:MergeItems, we need the revision IDs of the two edits,
since they’re passed into the “success” method.) It’s possible to
”forge” these success URLs and load them directly when no corresponding
edit actually took place; IMHO that’s fine and not something we need to
prevent.

Bug: T356149
Change-Id: I2eeca7e88360eac42ee14c18814287ea08d34fa6
Depends-On: If5cc2cf1d5985e08933d1477ba857a5372fe2d1e