Use EditEntity for ItemMergeInteractor
This means less duplicated code, and also makes dealing with IP Masking
and Temporary Users easier – we should still get the savedTempUser from
the returned status, but let’s do that separately.

Note that we still have to check permissions, because EditEntity only
checks “edit” permissions, not “merge” permissions, and that’s currently
not configurable (addRequiredPermission() was removed in change
Ibfb8b4c224, commit 2af5acfcef). I’d also be okay with making the
permission configurable in EditEntity again.

In the tests, I decided to use a real MediaWikiEditEntity rather than
mocking it, mainly because it means fewer changes to the test are
necessary – we can still use the same mocked permission checker and just
plug it into MediaWikiEditEntity, for instance. But I think it also
makes some sense to use the real service for the integration test here.

Bug: T356149
Bug: T356764
Change-Id: I919db202f9c733f01c6900bd1ebe00826bd85aca