Render expand/collapse of entitytermslist locally for anon users
EntityViewPlaceholderExpander currently renders the collapsed /
expanded state of the TermsList / TermBox based on a cookie for
users that are not logged in. This works locally but in production,
caching of the render means that the expand / collapse does not
work as expected over a page refresh for a non-logged-in or temp
user.

Here, we change the expand/collapse to be entirely client-side for
users that aren't able to set backend user options (i.e. Anonymous
and Temp users).

Bug: T351976
Change-Id: Id6bc261946462905bc7fc80a3c377df3c136c630