Make edit APIs return temp user redirect URL
Using this pattern when calling ResultBuilder:

    fn( $user ) => $this->getTempUserRedirectUrl( $params, $user )

I wasn’t sure how to make the getTempUserRedirectUrl() call happen at
first; this is the best version I could think of. The trait method is
protected, so if we passed the whole API module into the ResultBuilder,
the ResultBuilder would not be allowed to call it; also, most of the
addTempUser() callers already have the $params available in a local
variable, so using an arrow function also lets them pass it directly to
getTempUserRedirectUrl() without having to pass $params through
addTempUser().

Bug: T357024
Change-Id: I6a11a71143341ffbc34267cae8693aeb3a4c8ac2
Depends-On: I1c8b6b48a0e0ddb4cd10a0279f8f75ba0ab5afc5