Make EditEntity service return new context
For use cases like merging items and lexemes, where we make several
edits in a row, it’s important that we use a context with the new
temporary user for subsequent edits; I think it’s cleanest if EditEntity
creates and returns that context. (Setting the user of the original
context was discussed but advised against in WMF’s Slack [1].)

Also use this opportunity to consolidate and update the documentation of
the returned status value. The “new” field definitely isn’t being
returned anymore, and I’m not sure it ever was; but add the recently
added “savedTempUser” and the new “context”, and clarify that “revision”
is an EntityRevision, not any MediaWiki core type. Also, remove the
copies of the doc comments in MediaWikiEditEntity and only document the
methods on the interface, with updates from the slightly-diverged
MediaWikiEditEntity version and also a typo fix (toke → token ).

[1]: https://wikimedia.slack.com/archives/C01R06P8D1B/p1706705068528389?thread_ts=1706704546.420089&cid=C01R06P8D1B

Bug: T356149
Change-Id: Id0a54109b354bb7d840464ae7dffa61f225ef414