Introduce Status subclasses for EditEntity
TempUserStatus is a base class for several edit scenarios; EditEntity
returns EditEntityStatus, EntityRedirectCreationInteractor returns
EntityRedirectCreationStatus, ItemMergeInteractor returns
ItemMergeStatus. Each class has a type-safe factory method and getters.

For now, to preserve compatibility with existing code, we extend Status.
Note that Status is not currently marked as stable to extend, and it’s
unclear whether even StatusValue will be marked as stable to extend or
not; depending on the outcome of I6690fbbb13 / Ic1a8eccc53 / T358492, we
should either migrate to the StatusValue class later, or effectively
revert this and just use plain Status (or StatusValue?) again. That
said, I already based Gerrit changes for several other tasks on this
refactoring (before it became clear that there was a lot to discuss
about Status), and might not have much time to disentangle those changes
from this refactoring in the next few days, so for now we are taking on
this technical debt, at the risk of having to revert it later.

Note that WikibaseLexeme’s MediaWikiLexemeRedirector doesn’t use the
return value of EntityRedirectCreationInteractor::createRedirect() yet
(it will use it with change Id9ae7ce917), so we can change that without
a compatibility dance between the extensions.

Change-Id: I1b434034fc9def46eff222b865c003caf1846be7