Remove dead code in LinkTitles
As far as I can tell, Item is guaranteed to be non-null by the time we
call getAttemptSaveStatus() (perhaps it wasn’t in the past?), so let’s
remove the code that allowed it to be null. Otherwise – if the “fake”
Status from getAttemptSaveStatus() actually made it to buildResult() –
the code would crash anyways, since the added addTempUser() call from
change I6a11a71143 (commit 5549f82096) expects the value to be an array
with a 'savedTempUser' in it, not `true`.

Bug: T357024
Change-Id: Ib01b8c3e8e590fee9a461978724a77fedaedb078