Update termbox anonymous warning message for temp user case
Now that we are introducing temporary user support, it is no longer
the case that making an anonymous edit will expose the users' IP
addresses. As we have done with other components, we adjust the
warning message here to depend on the state of the repo where the
Termbox is running. In this patch, we update the termbox entrypoint
so that we can pass the tempuserenabled state of the hosting repo
into the termbox, and introduce new i18n messages.

Bug: T357404
Change-Id: Ib55211bad3054a4175c186d023343e8835cf9698