Process temp user authentication redirects in API edit responses
Termbox uses RepoApiWritingEntityRespository to save edits to the
repo. With the introduction of Temp Users, an API edit request by
an unauthenticated user may result in the creation of a temp user.
In this case, the Termbox needs to redirect to the redirecturl
returned from the API to complete the centralauth login for the
user.

Depends-On: I1b8389dc5cd5557e209a3c68b4fc3d38888466d1
Bug: T357151
Change-Id: I9370b7cc9e58254c55f4eb41ab7820e38359adce