Fix tests when IP Masking is enabled
IP Masking (Temporary Accounts / Temporary Users) will probably be
enabled in regular CI soon (T359043), and in any case we want to enable
it in secondary CI (T353961). Fix three tests that were failing.

WikiPageEntityStoreTest is obvious enough: it was explicitly using an
anonymous user, and bypasses the layer that’s responsible for
automatically creating temporary users (that’s in MediaWikiEditEntity).
Make it use a second user instead; the other user was already being
called a “sysop” in the comments, but AFAICT it wasn’t actually a sysop
(maybe getTestUser() used to default to a sysop, but currently it
doesn’t), so make it a sysop to distinguish it from the other user.
(Note that getTestUser() returns the same user when asked for the same
groups, so we definitely need *some* distinguishing groups between the
two test users.)

SpecialSetSiteLinkTest mainly needed to make sure that it updates the
existing $wgGroupPermissions instead of completely replacing them; in
order to automatically create temporary users, anons ('*') must have the
createaccount right. Fortunately setGroupPermissions() does the right
thing.

However, there’s also an unfortunate interaction between test cases:

1. addDBDataOnce() calls createItems(), which creates a test user and
   uses it to save edits. Because the test user is saved to the database
   and this happens during addDBDataOnce(), the 'user' table is marked
   to not be cleared between test runs.
2. testExecutePostModifySiteLink() runs and creates a first test user,
   which is numbered “1” according to the 'user_autocreate_serial'
   table. The 'user_autocreate_serial' table is marked as used.
3. MediaWikiIntegrationTestCase resets the 'user_autocreate_serial'
   table between test runs. (But not the 'user' table, because that one
   contains data from addDBDataOnce().)
4. testExecutePostRemoveSiteLink() runs and tries to create another test
   user. According to the reset 'user_autocreate_serial' table, it is
   numbered “1” again, but the previous “1” test user still exists
   because the 'user' table was not reset. The test fails.

Avoid this by explicitly touching the 'user_autocreate_serial' table in
addDBDataOnce(), so that it also doesn’t get reset between test runs.
This is ugly, but MediaWiki core doesn’t have a better solution yet.

Bug: T353961
Change-Id: Ib0a176fe335cf45e5b2cdd351c95d5da162a1eb2