Use names for test cases to prevent issues when numbering them
This has been possible in PHPUnit for a long time and some of the latest
testcases even already make use of it.

One test was moved to after rewriting them with sub-page syntax because
with subpage syntax the tests become identical regardless of whether
they have the empty string or null in place of the subpage.

Change-Id: Id0854e15c8f16ee9ec19b5d25eb7f22bf7c84cb1