Replace empty() with strict compares to '', 0 or null
Found usage of empty() on expression ... that appears to be always set.
empty() should only be used to suppress errors. See https://w.wiki/6paE

In most cases this change the behaviour of the string "0",
but from the code it seems unexpected that "0" was checked here.

Use of strict compare with int 0 change behaviour if null would be given

Bug: T358325
Change-Id: Ib27b1bfde8cbbf3c28173957496f230879bd6c93