REST: Allow "optional" path context for some errors
This is meant for errors that may require additional context depending
on the endpoint they're used with. For example, an INVALID_LABEL error
does not require a path when setting a single label, but when the
endpoint operates on the whole list of labels it's useful to know which
label is invalid.

Change-Id: Ia5c82ff473794c7e7807e5d5197204b19f8c9cac