Use relative sizes for icons
Currently the interface icons are defined in pixels, which means they
don't scale based on the user's default font size.

This changes most of them to be defined in rem instead,
using 16px = 1rem. As far as I can tell, that is the usual default.

There should be no visible changes if the default font size hasn't been
changed, but they should now scale up or down appropriately if it has.

Bug: T328387
Change-Id: I1f43cce24fe8f5d2e8c3e8a9940563d54fb40750