Add missing summaries for wbremovequalifiers, wbremovereferences
The -remove message is used when removing one qualifier or reference,
while the -update message is used when removing several (see
ChangeOps::apply()): in the former case, the summary will show further
details about the removed qualifier or reference, but not in the latter
case. Thanks to {{PLURAL}}, we can just use the same message for -remove
and -update in both cases.

Bug: T358975
Change-Id: Ie31f76e33a0c4883d9597bb3e3315c6e777a80aa