SimpleCacheWithBagOStuff: Remove double UTF-8 (1/3)
SimpleCacheWithBagOStuff uses utf8_encode() to convert the serialized
data from Latin 1 to UTF-8 before storing it in the cache. However, the
serialized data is actually UTF-8 already (guaranteed by MediaWiki core
since T137509 if not earlier, i.e. June 2016), so calling utf8_encode()
on it is pointless. (The function name is also confusing, which is why
it’s deprecated in PHP 8.2, see the PHP RFC [1].)

However, as this cache is used for some very high-volume data, we can’t
just change the data format all at once. Instead, we’ll roll out the
change in three steps: 1. enable SimpleCacheWithBagOStuff to handle
cached data in a new format, which doesn’t use the extra utf8_encode();
2. change it to emit this new format itself; and 3. remove support for
the old format. Each step will be rolled out in Wikimedia production in
three successive deployment trains. (The separation of step 1 and 2 is
necessary so that, if step 2 is deployed and then the train has to be
rolled back, Wikibase doesn’t crash when encountering cached data that
was stored by step 2.)

This commit represents step 1. The data in the cache can now be a
JSON-serialized array of two or three elements, and the array length
represents the format of the cached data: two elements means the old
format (HMAC, data), three elements means the new format (version, HMAC,
data). If we ever need to change the format again, we can just increment
the first element (version number).

[1]: https://wiki.php.net/rfc/remove_utf8_decode_and_utf8_encode

Bug: T324202
Change-Id: I61ab0823e61d693e6a5d7878e1d2918c35038d92
(cherry picked from commit 6a2de1e0d5d7574f17af40a93198b860996dfab5)