SimpleCacheWithBagOStuff: Remove double UTF-8 (2/3)
This is step 2 of the migration to remove the double encoding from
SimpleCacheWithBagOStuff (see change I61ab0823e6). The cache now emits
the new format, but can still read cached data that was stored in the
old format.

It turns out there is a test case (in SimpleCacheTestCase, which lives
in lib/tests/phpunit/ but was actually copied from an upstream project,
see T237164) that expects arbitrary binary strings to be supported by
the cache. As far as I can tell, we have no need for this (we always put
textual strings in the cache, and MediaWiki guarantees that the encoding
is UTF-8), so disable the test and instead add a test asserting that
non-UTF-8 binary data is not supported.

Bug: T324202
Change-Id: Ibb0c848808c43c37d27536bfdc6eaf9073e634a5