SimpleCacheWithBagOStuff: Remove double UTF-8 (3/3)
This is step 3 of the migration to remove the double encoding from
SimpleCacheWithBagOStuff (see change I61ab0823e6). The code is now
cleaned up and the cache only supports the new format.

The maximum cache duration I found (sharedCacheDuration in
$wgWBRepoSettings and $wgWBClientSettings) is just one day in Wikimedia
production (60 * 60 * 24 seconds), so this should be fine to deploy one
train after step 2.

Bug: T324202
Change-Id: I4e1df713a8a8fdfe4f070fcf6a5181e65d778ac7