REST: Rename old label and description validation classes
Labels and descriptions of an Item need to be validated against each
other, e.g. because they cannot be the same for a given language.
Thus, the existing ItemLabelValidator implementation performs a
database lookup to retrieve an Item's descriptions and the existing
ItemDescriptionsValidator looks up the Item's labels. Both is not
possible when creating a new Item, when labels and descriptions are
submitted together and cannot be looked up from the database.

These renamed validators would be deleted as soon as all the use cases depending on them use the new validators (those were created in If0fe2cc219c48b06d76dfd1179f7c3080acc8194).

Bug: T359848
Change-Id: I73b0984165e2545538e75f843a446b8adbd19dbf