SimpleCacheWithBagOStuff: Check cached data more thoroughly
Given that we don’t trust the cached data until we’ve validated the
HMAC, we shouldn’t just assume that it’s valid JSON of an array, or that
the signature or data to check are really strings. Check all those
things first, so we don’t get warnings or type errors (depending on PHP
version).

Bug: T324202
Change-Id: I414772b906dd491321504a9d002200d6674886d3