DataBridgeReferenceFormatter: Use message as text
Previously, the message would implicitly be stringified in the implode()
call below, which uses the “parse” format (see Message::__toString()). I
don’t think we want this format: all the other messages in this method
use ->text(); the method is documented (in its parent interface,
ReferenceFormatter) to return Wikitext, not HTML; and none of the
existing translations of wikibase-reference-formatter-snak-retrieved
need Wikitext processing anyways. Additionally, it looks like involving
the Parser here might cause problems in CI.

Bug: T362505
Change-Id: I7d04048b0ef84892c1a3920e5d2ec5e2c592892f