Align CacheAwarePropertyInfoStore caches with the lookup's
The PropertyInfoLookup service is a PropertyInfoTable wrapped in a
CachingPropertyInfoLookup using a WANObjectCache wrapped in a
CachingPropertyInfoLookup using a LocalServerObjectCache, while the
PropertyInfoStore service is a PropertyInfoTable wrapped *only* in a
CacheAwarePropertyInfoStore using a WANObjectCache.

It's the CacheAwarePropertyInfoStore's job to invalidate the cache if a
new property added, but since the PropertyInfoStore only knew about one
of the two, the other was left stale.

Change-Id: I6dc28447e4ba8a42366a08f3d6be26064bcfbbbb