Fix sampling of Lua function tracking
Per the attached task, we can't rely on Lua's
math.random as this is not properly seeded here.

Instead of implementing our own pseudo random
number generator here (or messing with the state
of the global Lua math random seed), we simply
track every nth function call (using a simple
counter and modulo). As we use an initial random
offset for this counter, the sampling should be
random enough for our use cases.

Bug: T360891
Change-Id: Id9a9ff29e68f02d161879047b5a88e9673ca5518