Support additional types in wbsearchentities
Let other extensions, like EntitySchema, register custom search helper
callbacks and support them from wbsearchentities. To make this possible,
allow the TermSearchResult entity ID to be null, and skip various steps
in that case. To ensure a consistent response structure, the
TermSearchResult’s metadata must in this case contain the keys that
would otherwise be added automatically by ConceptUriSearchHelper and
SearchEntities – id, title, pageid, url, and concepturi.

This is only supported in wbsearchentities. The other search API,
query+wbsearch (i.e., action=query with list=wbsearch or
generator=wbsearch, implemented in QuerySearchEntities), relies even
more on being able to look up a title for an entity ID than
wbsearchentities does; while we could get that title from the title or
pageid in the metadata, that feels ugly to me. So the type= parameter of
that API is still limited to “real” (registered) entity types, and
calling it with type=entityschema just yields an “unrecognized value for
parameter” error.

Bug: T362955
Change-Id: Ie530ce2816be07e50d441687e39bda7d3ab6e3e0