REST: Split labels/descriptions validation in two
This models validation of labels and descriptions in two parts: 1.
validating the syntax/structure, 2. validating their contents. This
allows us to work around the chicken/egg problem of needing the labels
to fully validate the descriptions and vice versa.

The new validators are also more reusable for patch use cases, which
will be adjusted in later patches.

Bug: T362606
Change-Id: Id447fec135ea0d4ae12a0f46d99ac55bac52d09f