Replace wikibase-monolingualtext message with hard-coded HTML
No language actually customized this message, and expecting them to
duplicate the full HTML markup is silly anyways. Instead, hard-code the
HTML structure in MonolingualHtmlFormatter, and use MediaWiki’s existing
messages to localize the language-dependent element: the space between
text and language name, and the parentheses around the language name.
(In e.g. Chinese, the space is blank and the parentheses are fullwidth.)

Bug: T361005
Change-Id: I0b98a79d308a1d1fc648cf25ad4445ba4adf06fb