Clean up PropertyDataTypeLookup service wiring
The BaseDataModelDeserializerFactory service wiring was previously using
a separate PropertyDataTypeLookup; this was necessary to avoid a
circular dependency in the client services:

> Recursive service instantiation: Circular dependency when creating
> service! WikibaseClient.RestrictedEntityLookup ->
> WikibaseClient.EntityLookup -> WikibaseClient.Store ->
> WikibaseClient.WikibaseServices ->
> WikibaseClient.BaseDataModelDeserializerFactory ->
> WikibaseClient.PropertyDataTypeLookup -> WikibaseClient.EntityLookup

We can avoid this by making the PropertyDataTypeLookup wiring a bit more
lazy; by letting PropertyInfoDataTypeLookup take a callable as the
fallback lookup, which is only called when it’s first needed, we can
avoid PropertyDataTypeLookup directly needing EntityLookup. This in turn
lets us use the normal PropertyDataTypeLookup service when creating the
BaseDataModelDeserializerFactory.

In repo, this is not necessary, because the creation of the whole
PropertyInfoDataTypeLookup is effectively lazy already (it happens in
the callback returned by PropertyServices::getServiceDefinitions());
however, we can improve that code by injecting the correct $services
instance instead of just relying on the global default services.

Bug: T362004
Change-Id: I031c34827e140483897c2ded3c7ef62f4b727aaa