Register parsers for both data types and value types
This restores registering value type specific parsers, which was removed
in I635e8119350fed007120246ece1d11bfdafa11b1. We identified one place
(badge editing) that required the value type parser fallback. Even
though that particular case could be refactored to explicitly use an
item value parser, we're unsure whether there are other places relying
on the fallback, so it seems best to restore this for now and
investigate later.

Bug: T366236
Bug: T366275
Change-Id: I44c0fb8712a5799ac24711b2dc0ff3ef3a11a667