Add 'deserializer-builder' to DataTypeDefinitions
I690780ee89da4625de302302709511c72cb9d89a enabled data type specific
value parsing/deserializing and used the parser for both. We now
realized that parsing and deserializing need to be separated, so this
patch adds a deserializer-builder field to DataTypeDefinition. For now,
this is only used for the existing generic data value deserializers. A
later patch will add a `SnakValueDeserializer` and use that in places
that currently use `SnakValueParser` when they should be using the
deserializer.

Bug: T365794
Change-Id: I669b486e61be66fbb46e1cd609e8400a474d0ffd