REST: Rename "FullEntityEditSummary" -> "WholeEntityEditSummary"
While "full" can mean "not lacking or omitting anything; complete" this word, to me, has a stronger connotation with "containing or holding as much or as many as possible; having no empty space" which isn't the meaning we want to convey with this name.

"Whole" has a strong connotation with "all of; entire" which is the meaning we want with less ambiguity.

Change-Id: I5dbe399913a24760f2bcd36b063ab18171ae4e70