REST: Use term type specific validators
We previously used a generic LanguageCodeValidator in all cases, which
doesn't work, since e.g. "mul" is a valid alias and label language, but
cannot be used for descriptions. This creates separate interfaces for
each type and implements them all using the same implementation which is
configured using type specific validators in our service wiring.

Bug: T367070
Change-Id: Iad7ff34a44424ad18f30e449c287c0ce2fe10470