Remove deprecated `setContentLang` call
The `setUp` for `SpecialAvailableBadgesTest` switches the language
to `qqx`. There is no `tearDown` to restore the previous language
and the test seems to pass okay without the language change.

In addition, because of the current implementation of
`Title::isMainPage`, setting the content language here causes a
`qqx`-language main page to be cached in the static field of
`Title` which interacts badly with subsequent tests.

Bug: T368369
Change-Id: Iab57ac7be8e2beb9ba69fc1bf5d64484dfd2c910