PropertyValueExpertsModule: Turn on enableModuleContentVersion()
This should ensure that users see the new version of this module when
the contents changed (e.g. when a new data type is added). We already do
the same in some other modules (MediaWikiConfigModule, SitesModuleBase),
and I think the runtime cost should be okay: calling getScript() on a
PropertyValueExpertsModule seems to take about 20.8 microseconds in
local testing.

Bug: T369155
Change-Id: I6169154d2f15d0895ec51093f18cf202fe32d189