Defs::getSearchIndexDataFormatterCallbacks(): Support $mode
Make DataTypeDefinitions’ getSearchIndexDataFormatterCallbacks() support
either “prefixed” or “resolved” mode, like the other methods in this
class. For consistency, also make “resolved” the default; this is a
breaking change, but the result of this method is only ever used in
WikibaseCirrusSearch’s StatementsField, which is ready for the change
(see Depends-On), so that should be fine.

The search index data formatter callbacks are different to the other
callbacks in that they’re partial / incomplete: for example,
PT:globe-coordinate / VT:globecoordinate are missing a search index data
formatter (I assume you’re not supposed to add 'globe-coordinate' to the
$wgWBRepoSettings['searchIndexTypes']). Add a new flag to applyMode()
and resolveValueTypeFallback() to accommodate this and not throw an
error for these callbacks.

Bug: T369149
Change-Id: Ie44ad9f5d81105c2de903d6430ae379bf8b7767e
Depends-On: I62d24b0388e3ed1eaac9b0faf6c75f7ebb9c5d46