Add setup for EntitySchema to Wikibase CI setup
We want to be able to test EntitySchema in client-only mode.
Currently, there is an existing client-only mode CI test job -
`wikibase-client-php81` - that only runs for Wikibase patches.
With I93a08d9e8c88aacd2ec4dce8a052343fbf365a13, we enable those
jobs for EntitySchema patches. In order to setup EntitySchema
correctly in that case, we need to enable the datatype via the
global environment setting and configure `wgEntitySchemaIsRepo`
appropriately.

This change should have no effect for the existing Wikibase checks
(and if it does, that would be a bug).

Bug: T367156
Change-Id: I73f465d86eafc377ad69c5054404673108c3f182