Validate language code in wbformatvalue
This is a special case of T323778: The "lang" option, being the most
commonly used across many different formatters, has the greatest
potential for breakage if invalid values are specified in it.
Previously, we would try to put any language code from the user into the
FormatterOptions and format the value with that (although language codes
that didn’t pass LanguageNameUtils::isValidCode() – a weaker check than
isValidBuiltInCode() – would crash in the LanguageFallbackChainFactory
via OutputFormatValueFormatterFactory::applyLanguageDefaults() and
therefore return a param-illegal error). Now, we validate the language
and set it to "und" if it’s not valid; this matches what
EntityHandler::getValidUserLanguage() does. If the value in the options
isn’t even a string, then we raise an API error (previously, I assume
this would have resulted in an internal TypeError sooner or later).

The `switch` in setValidOption() is meant to be extensible for other
options later, if we want to make this API validate the options in
general (per T323778#9498463); but IMHO validating the language is
reasonable even if we otherwise don’t go with this approach, because
this option is special (it’s the only one defined in a ValueFormatter
constant) and already special-cased in EntityHandler (as mentioned
above).

Bug: T323778
Change-Id: Icaec1e3b8ee41520c0cff1a20ca172e5506ccfda