Require nonempty term types in APIs
If the term types are empty, the underlying database query can’t be
built (SQLPlatform::makeList: empty input for field wbtl_type_id).
Let’s make the API enforce that the types are nonempty and throw an
error otherwise, using code based on ApiQueryUserContribs in core.

I tried also adding an assertion that the types must be nonempty in
PrefetchingEntityTermLookupBase::prefetchTerms(), but that caused other
issues and I’m not sure it’s worth fixing them; the underlying
SQLPlatform error is probably enough, to be honest.

Bug: T370769
Change-Id: Ic10386e98f94b9864994d545371d5ba0902fd4d8