Relax edit summary length limit for undo actions
Since T6714, the edit summary length limit has been increased
to 500. Update the form used for undo (and restore) actions
to allow as many characters as possible.

Note that the custom edit summary is appended to a system-generated
autocomment, which also counts towards the limit. Therefore, we
generate the autocomment prior to showing the form and subtract
its length from the limit. This makes the patch somewhat more
involved.

Bug: T308859
Change-Id: Ie752545db475be353daf850fadb599b9f0ed6692