Create temp user before checking edit filters
This is the new recommendation per T359405 (see discussion in T356148) –
if the edit filter results in a log entry, that log entry needs a
temporary user.

In ItemMergeInteractorTest, we need to change the way the edit filter
detects the user to block. The previous code actually resulted in a user
with our custom name but user ID 0; for the temp user creator, this
counted as anonymous, so the edit filter would see a temp user instead.
Use a real user via the test user registry, and since we can’t set the
user name there without some hassle, use a custom group instead.

Bug: T371085
Change-Id: I8342b392f35f5c9d2b293b83f12b7b7fe771d019